requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
llama-index-core>=0.10.0,<0.11.0
llama-index>=0.10.0,<0.11.0
llama-index-vector-stores-faiss>=0.1.0,<0.2.0
llama-index-llms-openai>=0.1.0,<0.2.0
llama-index-embeddings-voyageai>=0.1.0,<0.2.0
psycopg2-binary>=2.9.0
pydantic>=2.5.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.0

